Fix most spline-related issues
