+++ jetson/ra_kinematics/test/motion_planner_test_build.txt
@@ -2,6 +2,6 @@ project('ra_kinematics', 'cpp', default_options : ['cpp_std=c++14'])
 
 liblcm = dependency('lcm')
 
-executable('jetson_ra_kinematics', 'test/motion_planner_tests.cpp', 'arm_state.cpp', 'kinematics.cpp', 'motion_planner.cpp', 'utils.cpp',
+executable('jetson_ra_kinematics', 'test/motion_planner_tests.cpp', 'arm_state.cpp', 'kinematics.cpp', 'motion_planner.cpp', 'utils.cpp', 'spline.cpp',
            dependencies : [liblcm],
            install : true)
